05-Text2SQL-Evaluation/requirements.txt
pymysql>=1.1.0
sqlalchemy>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
python-dotenv>=1.0.0
transformers>=4.30.0
openai>=1.0.0
pyyaml>=6.0.0
torch>=2.0.0 